vllm_client/requirements.txt
openai>=1.35.0
pillow>=10.0.0
soundfile>=0.12.1
numpy>=1.24.0
python-dotenv>=1.0.0
